refactor: improve Search component performance and responsiveness
